+++ schema/src/main/scala/org/constellation/schema/serialization/SchemaKryoRegistrar.scala
@@ -100,9 +100,11 @@ object SchemaKryoRegistrar
         (Class.forName("scala.collection.immutable.Nil$"), 116),
         (Class.forName("scala.collection.immutable.Map$EmptyMap$"), 136),
         (classOf[SnapshotProposalPayload], 200),
-        (classOf[Signed[SnapshotProposal]], 201),
+        (classOf[Signed[_]], 201),
         (classOf[SnapshotProposal], 202),
         (classOf[MajorityInfo], 203),
-        (classOf[HeightRange], 204)
+        (classOf[HeightRange], 204),
+        (classOf[CheckpointBlockPayload], 205),
+        (classOf[FinishedCheckpointBlock], 206)
       )
     ) {}