+++ src/main/scala/org/constellation/domain/p2p/client/CheckpointClientAlgebra.scala
@@ -1,8 +1,9 @@
 package org.constellation.domain.p2p.client
 
+import org.constellation.gossip.state.GossipMessage
 import org.constellation.infrastructure.p2p.PeerResponse.PeerResponse
 import org.constellation.schema.GenesisObservation
-import org.constellation.schema.checkpoint.{CheckpointCache, FinishedCheckpoint}
+import org.constellation.schema.checkpoint.{CheckpointBlockPayload, CheckpointCache, FinishedCheckpoint}
 import org.constellation.schema.signature.{SignatureRequest, SignatureResponse}
 
 trait CheckpointClientAlgebra[F[_]] {
@@ -10,7 +11,11 @@ trait CheckpointClientAlgebra[F[_]] {
 
   def getCheckpoint(hash: String): PeerResponse[F, Option[CheckpointCache]]
 
+  def checkFinishedCheckpoint(hash: String): PeerResponse[F, Option[CheckpointCache]]
+
   def requestBlockSignature(signatureRequest: SignatureRequest): PeerResponse[F, SignatureResponse]
 
   def sendFinishedCheckpoint(checkpoint: FinishedCheckpoint): PeerResponse[F, Boolean]
+
+  def postFinishedCheckpoint(message: GossipMessage[CheckpointBlockPayload]): PeerResponse[F, Unit]
 }